Update Final_Assesment2.Rmd
- changes (same as I wrote in whatsapp):

Title page:
- it says in the requirements for the title page to also put our university. I copied over the part containing the information on the group members from the first assignment and checked that the information is correct.

Task 1:
- wording suggestion: First, we chose three variables representing relevant social characteristics given above theory that are stored in the variables "education", "marital status", and "wealth".
- wording / clarity suggestion: We then used forward stepwise selection to identify those covariates with the highest empirical relevance for the dependent variable "rejection". From the resulting list of nine empirically highly relevant variables, we picked four that are also theoretically relevant, thus obtaining a total of seven regressors. 
- moved formula for conditional probability for average individual to the second phrase, because it is referring to the 9.04% and the x-bar would otherwise be wrong in sentence before as that was referring to the general case for any individual
- spelling: "bankruptcy"
- capitalization of column headers in table of regressors

Tasks 2 & 3: only a few minor changes to the wording and some spelling in Task 3. Also in task 3, percentages are given for the accuracy of predicted positives and negatives and these are described correctly but not named to be "sensitivity" or "specificity" which I think we should probably do. So I added that in one of the sentences like this: "We observe some improvement in the accuracy of predicted positives (sensitivity now at 36%) while keeping the specificity at a very high level (now 96%)." This should be correct but I haven't eaten and I'm tired so please double check if that makes sense so that I'm not accidentally mixing it up here xD

4: went over it one more time and corrected a typo

So as you can see, it's only little changes to the wording. The whole thing should now be a few words shorter than before.